reseting default to bfloat16

+++ berkeley-function-call-leaderboard/bfcl_eval/model_handler/local_inference/smollm2.py
@@ -11,10 +11,10 @@ class SmolLm2Handler(OSSHandler):
         temperature,
         registry_name,
         is_fc_model,
-        dtype="half",
+        dtype="bfloat16",
         **kwargs,
     ) -> None:
-        super().__init__(model_name, temperature, registry_name, is_fc_model,model_max_len=81920 ,dtype = dtype ,**kwargs)
+        super().__init__(model_name, temperature, registry_name, is_fc_model, dtype=dtype, model_max_len=81920, **kwargs)
     @override
     def _format_prompt(self, messages, function):
         """
